+++ tests/testthat/test-occumbData.R
@@ -3,16 +3,16 @@
 
 test_that("Mode check for y works", {
     expect_error(new("occumbData", y = array(list(), dim = rep(2, 3))),
-                 "'y' should be a 3D-array of integers, not lists.")
+                 "Elements of 'y' are lists but should be integers")
 })
 
 test_that("Dimension check for y works", {
     expect_error(new("occumbData", y = array(1:4, dim = rep(2, 1))),
-                 "'y' should be a 3D-array.")
+                 "'y' is not a 3D-array")
     expect_error(new("occumbData", y = array(1:4, dim = rep(2, 2))),
-                 "'y' should be a 3D-array.")
+                 "'y' is not a 3D-array")
     expect_error(new("occumbData", y = array(1:16, dim = rep(2, 4))),
-                 "'y' should be a 3D-array.")
+                 "'y' is not a 3D-array")
 })
 
 test_that("Check for missing values for y works", {
@@ -22,14 +22,14 @@ test_that("Check for missing values for y works", {
 
 test_that("Integer check for y works", {
     expect_error(new("occumbData", y = array(1:8 + 0.1, dim = rep(2, 3))),
-                 "'y' contains non-integer value.")
+                 "'y' contains non-integer value")
     expect_error(new("occumbData", y = array(c(Inf, 1:7), dim = rep(2, 3))),
-                 "'y' contains non-integer value.")
+                 "'y' contains non-integer value")
 })
 
 test_that("Check for non-negative values for y works", {
     expect_error(new("occumbData", y = array(c(-1, 1:7), dim = rep(2, 3))),
-                 "'y' contains negative value.")
+                 "'y' contains negative value")
 })
 
 ## Tests for covariates
@@ -38,27 +38,27 @@ test_that("Check for covariate names works", {
     names(unnamed_list2) <- c("a", "")
     expect_error(new("occumbData", y = array(1:8, dim = rep(2, 3)),
                                    spec_cov = unnamed_list1),
-                 "'spec_cov' contains unnamed element.")
+                 "'spec_cov' contains unnamed element")
     expect_error(new("occumbData", y = array(1:8, dim = rep(2, 3)),
                                    spec_cov = unnamed_list2),
-                 "'spec_cov' contains unnamed element.")
+                 "'spec_cov' contains unnamed element")
     unnamed_list1 <- unnamed_list2 <- list(rep(1, 3), rep(1, 3))
     names(unnamed_list2) <- c("a", "")
     expect_error(new("occumbData", y = array(1:8, dim = rep(2, 3)),
                                    site_cov = unnamed_list1),
-                 "'site_cov' contains unnamed element.")
+                 "'site_cov' contains unnamed element")
     expect_error(new("occumbData", y = array(1:8, dim = rep(2, 3)),
                                    site_cov = unnamed_list2),
-                 "'site_cov' contains unnamed element.")
+                 "'site_cov' contains unnamed element")
     unnamed_list1 <- unnamed_list2 <-
         list(matrix(1:2, 1, 2), matrix(1:2, 1, 2))
     names(unnamed_list2) <- c("a", "")
     expect_error(new("occumbData", y = array(1:8, dim = rep(2, 3)),
                                    repl_cov = unnamed_list1),
-                 "'repl_cov' contains unnamed element.")
+                 "'repl_cov' contains unnamed element")
     expect_error(new("occumbData", y = array(1:8, dim = rep(2, 3)),
                                    repl_cov = unnamed_list2),
-                 "'repl_cov' contains unnamed element.")
+                 "'repl_cov' contains unnamed element")
 })
 
 test_that("Check for covariate name overlap works", {
@@ -91,56 +91,56 @@ test_that("Check for covariate name overlap works", {
 test_that("Dimension check for covariates works", {
     expect_error(new("occumbData", y = array(1:8, dim = rep(2, 3)),
                                    spec_cov = list(a = rep(1, 1))),
-                 "Length of 'a' should match the number of species.")
+        "'a' must have a length equal to the number of species")
     expect_error(new("occumbData", y = array(1:8, dim = rep(2, 3)),
                                    spec_cov = list(a = rep(1, 1),
                                                    b = rep(1, 3))),
-        "Length of 'a' and 'b' should match the number of species.")
+        "'a' and 'b' must have a length equal to the number of species")
     expect_error(new("occumbData", y = array(1:8, dim = rep(2, 3)),
                                    site_cov = list(a = rep(1, 1))),
-                 "Length of 'a' should match the number of sites.")
+        "'a' must have a length equal to the number of sites")
     expect_error(new("occumbData", y = array(1:8, dim = rep(2, 3)),
                                    site_cov = list(a = rep(1, 1),
                                                    b = rep(1, 3))),
-        "Length of 'a' and 'b' should match the number of sites.")
+        "'a' and 'b' must have a length equal to the number of sites")
     expect_error(new("occumbData", y = array(1:8, dim = rep(2, 3)),
                                    repl_cov = list(a = rep(1, 1))),
-                 "'a' should be a matrix with J rows and K columns.")
+        "'a' must have a number of rows equal to the number of species and a number of columns equal to the number of sites")
     expect_error(new("occumbData", y = array(1:8, dim = rep(2, 3)),
                                    repl_cov = list(a = rep(1, 1),
                                                    b = rep(1, 3))),
-        "'a' and 'b' should be a matrix with J rows and K columns.")
+        "'a' and 'b' must have a number of rows equal to the number of species and a number of columns equal to the number of sites")
     expect_error(new("occumbData", y = array(1:8, dim = rep(2, 3)),
                                    repl_cov = list(a = matrix(1:2, 1, 2))),
-                 "'a' should be a matrix with J rows and K columns.")
+                 "'a' must have a number of rows equal to the number of species and a number of columns equal to the number of sites")
     expect_error(new("occumbData", y = array(1:8, dim = rep(2, 3)),
                                    repl_cov = list(a = matrix(1:2, 1, 2),
                                                    b = matrix(1:2, 2, 1))),
-        "'a' and 'b' should be a matrix with J rows and K columns.")
+        "'a' and 'b' must have a number of rows equal to the number of species and a number of columns equal to the number of sites")
 })
 
 test_that("Check for missing covariate values works", {
     expect_error(new("occumbData", y = array(1:8, dim = rep(2, 3)),
                                    spec_cov = list(a = c(1, NA))),
-                 "'spec_cov' contains missing value.")
+                 "'spec_cov' contains missing value")
     expect_error(new("occumbData", y = array(1:8, dim = rep(2, 3)),
                                    site_cov = list(b = c(1, NA))),
-                 "'site_cov' contains missing value.")
+                 "'site_cov' contains missing value")
     expect_error(new("occumbData", y = array(1:8, dim = rep(2, 3)),
                                    repl_cov = list(c = matrix(c(1:3, NA), 2, 2))),
-                 "'repl_cov' contains missing value.")
+                 "'repl_cov' contains missing value")
 })
 
 test_that("Check for infinite covariate values works", {
     expect_error(new("occumbData", y = array(1:8, dim = rep(2, 3)),
                                    spec_cov = list(a = c(1, Inf))),
-                 "'spec_cov' contains infinite value.")
+                 "'spec_cov' contains infinite value")
     expect_error(new("occumbData", y = array(1:8, dim = rep(2, 3)),
                                    site_cov = list(b = c(1, Inf))),
-                 "'site_cov' contains infinite value.")
+                 "'site_cov' contains infinite value")
     expect_error(new("occumbData", y = array(1:8, dim = rep(2, 3)),
                                    repl_cov = list(c = matrix(c(1:3, Inf), 2, 2))),
-                 "'repl_cov' contains infinite value.")
+                 "'repl_cov' contains infinite value")
 })
 
 ### Test for check_covariate_mode() --------------------------------------------